Testing enki and CrossWidnow
CrossWindows don't go well with enkiTS, the overhead and wait is bigger than actual execution.

Parallellism is only beneficial when there is careful planification and safe data to pair with. Windows don't have that much events that incurr a high cost of handling to get value from enkiTS approach to parallellism.

Maybe with a more naive approach to set each window to a task and just keep sending it, then it might just work.
- Write own more direct task queue, which might be tremendously unsafe or just do parallelism when actually needed for not trivial tasks?