Multi queue scheduling and Dependencies
Right now this is actually slower than single queue I think, will add to a branch, test back with a single queue.

For what It seems, at 24threads there is a 10x perf increase with this very bad and basic test. Cost of adding a task is too high at 4us, keep adding tasks, entities and things like that and it will snowball a lot the total time taken really to actually run tasks

There is a ginormouse issue is at checking dependencies. Just checking on one dependency averages a cost of 12us which will increase dramatically with more dependencies. It goes through all threads and checks on each one if the task is being ran or is held there, on a dependency which is not found right now so not even part of a missed task entry and having to take another...

Solution to dependency check would be for the dependency vector to already know at which thread is being held onto, which could be part of a new class called TaskFinder{TaskID, thread_index}
- Would limit search  to a single thread instead of 24
- With task stealing between threads, the children ought to be saved too and then update their dependencies...