Cost of scheduling is not acceptable
cost of scheduling a task is not really acceptable, even if we leave it as it is, spacing out when the ordering changes does not really solve the issue.

On debug the average cost of schedule + check dependency is too high at 4us each~, 3us total in release, still at a lot of entities this becomes bad, like limitingly bad that could hold max around 60k tasks per frame (not entities, tasks, without taking into account the preparation for each task).

I don't think it is something I should be thinking about right now, next step is to implement the thread at which to check the dependency and task stealing.