Vulkan linking
Had some issues with linking, Vulkan_LIBRARY and their subsequent linkage did not work. Have to add both local and environment Vulkan_LIBRARY for it to work and find the package. After that it just has to link to the folder.

I am mixed in how to make the modules - completely detached and the main app interacts with them or are they together with similar interfaces?